@@ -1,4 +1,4 @@
 ﻿global using CPUFramework;
 global using System.Data;
-global using System.Data.SqlClient;
+global using Microsoft.Data.SqlClient;
 global using System.Diagnostics;
